requirements.txt
lxml>=6.0.1
python-dotenv>=1.1.1
requests>=2.32.5
PySide6>=6.9.1
beautifulsoup4>=4.13.5
kivy>=2.3.1
buildozer>=1.5.0
cython>=3.1.3